+++ packages/core/src/index.ts
@@ -1,2 +1,2 @@
 export { createCoreRuntimeStatus, type CoreRuntimeStatus } from "./runtime";
-
+export * from "./torrent/index";